// apps/web/src/lib/generated/models/EnterpriseAuditPageRead.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { EnterpriseAuditEventRead } from './EnterpriseAuditEventRead';
export type EnterpriseAuditPageRead = {
  items: Array<EnterpriseAuditEventRead>;
  next_cursor: (string | null);
};