// apps/web/src/lib/generated/models/UserGroupRead.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type UserGroupRead = {
  active: boolean;
  created_at: string;
  description: string;
  id: string;
  member_count: number;
  member_ids: Array<string>;
  name: string;
  tenant_id: string;
  updated_at: string;
  version: number;
};